LLMBackEnd/requirements.txt
pydantic==1.10.13
langchain==0.1.16
langchain-core==0.1.45
langchain-openai==0.0.6
langgraph==0.0.17
fastapi==0.115.12
uvicorn==0.33.0
python-dotenv
